+++ js-miniapp-sdk/CHANGELOG.md
@@ -1,6 +1,6 @@
 ## CHANGELOG
 
-### 1.20.0 (2024-xx-xx)
+### 1.20.0 (2024-03-21)
 - **Feature:** Added new interfaces `set(key: string, value: string)`, `get(key: string)`, `remove(key: string)`, `clearMiniAppPreferences()` which uses the native storage features like Shared Preferences/User defaults to store anything from MiniApp.
 - **Feature:** Updated HostEnvironmentInfo to have `hostBuildType`, `deviceToken` and `pushToken`
 - **Feature:** Added new interface `miniAppFinishedLoading()` which can be used by miniapp to notify the host app that it has finished loading.